+++ Objective-Zip/OZFileInZipInfo.h
@@ -43,10 +43,10 @@
 #pragma mark Properties
 
 @property (nonatomic, readonly) NSString *name;
-@property (nonatomic, readonly) NSUInteger length;
+@property (nonatomic, readonly) unsigned long long length;
 @property (nonatomic, readonly) OZZipCompressionLevel level;
 @property (nonatomic, readonly) BOOL crypted;
-@property (nonatomic, readonly) NSUInteger size;
+@property (nonatomic, readonly) unsigned long long size;
 @property (nonatomic, readonly) NSDate *date;
 @property (nonatomic, readonly) NSUInteger crc32;
